hangman: fix for segfaulting
Added a trim_right_space to fix a weird bug where if I edited the
word_list.txt file, then ran hangman in debug mode with './hangman test'
it would segfault. Even if nothing was added/removed from word_list.txt.
Originally, I made most of this in Windows. So I thought it might be a
LF/CRLF issue. But everything I've read says git and odin take those
kind of things into account. So I don't know what was causing it, but
this seems to fix it.